[OU-FIX] calendar: Reset sequence field in resource.calendar.attendance
As the field `sequence` exists previously in 16, and you were able to
change it on screen:

![imagen](https://github.com/user-attachments/assets/7201233f-de3e-465c-9bca-df0c866a7270)

there's a chance that your DB have some sequence values that alters the
previous order when migrating.

Thus, the safest thing is to reset the sequence value to its default
value for preserving always the previous order.

TT49327