[OU-FIX] account: Rename XML-IDs for installed CoA elements
On v17, when you load a CoA into a company, the CoA elements are still
given an XML-ID with the company ID + `_` + the original template XML-ID,
but now, instead of putting the module containing the template, all of
them are put with the module `account`. Reference:

https://github.com/odoo/odoo/blob/b9abe46c1492b09e369434e76ec8196c6b02dd19/addons/account/models/chart_template.py#L608

Thus, we need to rename the module for all the existing CoA elements
XML-IDs with this pattern.